fix: 코드 전체적인 수정
* CSS 수정
* SideBar 클릭 시 하이라이트 안되는 에러 수정
* generateTextIndent : 들여쓰기
* grenerateTitle : title 값이 없으면 UNTITLED 값이 나오도록

+++ src/App.js
@@ -1,9 +1,9 @@
 import DocumentEditPage from "./components/DocumentEditPage/DocumentEditPage.js";
 import SidebarContainer from "./components/Sidebar/SidebarContainer.js";
 import { fetchDocuments } from "./utils/api.js";
-import { DEFAULT_DOCUMENT_ID, DOCUMENTS_ROUTE, NEW, NEW_PARENT, OPENED_ITEMS } from "./utils/constants.js";
+import { FIRST_DOCUMENT_ID, DOCUMENTS_ROUTE, NEW, NEW_PARENT, OPENED_ITEMS } from "./utils/constants.js";
 import { initRouter, push } from "./utils/router.js";
-import { setItem, getItem } from "./utils/storage.js";
+import { setItem, getItem, removeItem } from "./utils/storage.js";
 import { setDocumentTitle } from "./utils/validation.js";
 
 export default function App({ $target }) {
@@ -15,7 +15,7 @@ export default function App({ $target }) {
     const createdDocument = await fetchDocuments("", {
       method: "POST",
       body: JSON.stringify({
-        title: document.title,
+        title: "",
         parent: getItem(NEW_PARENT, null),
       }),
     });
@@ -28,25 +28,28 @@ export default function App({ $target }) {
     });
   };
 
-  const onDelete = async (documentId) => {
+  const onDelete = async (currentDocumentId, removedDocumentId) => {
+    if (removedDocumentId === FIRST_DOCUMENT_ID) {
+      alert("첫 페이지는 지우지 말아주세요 :D");
+      return;
+    }
     if (!confirm("페이지를 삭제하시겠습니까?")) return;
 
-    await fetchDocuments(documentId, {
+    await fetchDocuments(removedDocumentId, {
       method: "DELETE",
     });
 
     const openedItems = getItem(OPENED_ITEMS, []);
-    const index = openedItems.indexOf(documentId);
-    if (index > -1) {
-      setItem(OPENED_ITEMS, [...openedItems.slice(0, index), ...openedItems.slice(index + 1)]);
-    }
+    setItem(
+      OPENED_ITEMS,
+      openedItems.filter((item) => item !== removedDocumentId)
+    );
 
-    const currentId = documentEditPage.state.documentId;
-    if (currentId === documentId) {
-      documentEditPage.setState({ documentId: DEFAULT_DOCUMENT_ID });
-      push(`${DOCUMENTS_ROUTE}/${DEFAULT_DOCUMENT_ID}`);
+    if (currentDocumentId === removedDocumentId) {
+      documentEditPage.setState({ documentId: FIRST_DOCUMENT_ID });
+      push(`${DOCUMENTS_ROUTE}/${FIRST_DOCUMENT_ID}`);
     } else {
-      documentEditPage.setState({ documentId: currentId });
+      documentEditPage.setState({ documentId: currentDocumentId });
     }
 
     sidebarContainer.render();
@@ -106,7 +109,7 @@ export default function App({ $target }) {
       documentId: isNaN(documentId) ? documentId : parseInt(documentId),
     });
 
-    if (isNaN(documentId)) {
+    if (!isNaN(documentId)) {
       sidebarContainer.setState({
         selectedId: parseInt(documentId),
       });
